(ref) port recent 0.16.0 updates
