--- recipe/patches/xxxx-remove-langref-std.txt
@@ -1,8 +1,0 @@
-test_comptime_unwrap_null.zig
-sentinel-terminated_pointer.zig
-test_variadic_function.zig
-cImport_builtin.zig
-verbose_cimport_flag.zig
-test_coerce_tuples_arrays.zig
-libc_export_entry_point.zig
-runtime_index_out_of_bounds.zig
